Attempted repair of PR1814
verilog.js fixed minor bug, but unable to PR1814
logix.js added tag in CircuitElement.prototype.generateVerilog so it takes tag - because unable to incorporate verilog.js, unable to use PR#1814
module.js has minor changes to write tag, otherwise same as PR#1814
Counter, sequential, subcircuit incorporated most changes in PR#1814

Problems:
Clock is translated as input, PR#1814 does not do this
Splitter is very hard to get to work properly. I used a lot of trial and error with verilog.js and logix.js. The change in code from PR#1814 made these changes not work.

Specific errors:
Splitter inputs need to be a single variable with bitwidth, sometimes this is to the case
Splitter output sometimes disappear completely, because of DFS from input of module, and sometimes splitter input is connected to an input of device or module
Making Clock an Input of the module sometimes resulted in it being an output or wire

I think the Splitter and Clock logic needs to be reworked so PR#1814 can work well
I have tried to incorporate as much of the changes as I can
Thanks,
James

Fixes #

Describe the changes you have made in this PR -
Screenshots of the changes (If any) -
Note: Please check Allow edits from maintainers. if you would like us to assist in the PR.